Move last memcache bits out of dynamo_collector
